# ruian-swagger/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: SYSNET RUIAN services API
  description: SYSNET RÚIAN service REST API
  contact:
    email: info@sysnet.cz
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: 1.0.2
servers:
- url: https://virtserver.swaggerhub.com/sysnetcz/RUIAN/1.0.2
...
paths:
  /reverse/parcel/jtsk:
    get:
      tags:
      - public
      summary: find land parcel by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ plot of land\n"
      operationId: parcela_api
      parameters:
      - name: x
        in: query
        description: JTSK x coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -770559.237
      - name: "y"
        in: query
        description: JTSK y coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -1068317.864
      responses:
        "200":
          description: plot (of land) object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcela'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find plot (of land) by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ plot of land\n"
      operationId: parcela_post_api
      requestBody:
        description: JTSK coordinates
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jtsk'
      responses:
        "200":
          description: plot (of land) object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcela'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/parcel/wgs:
    get:
      tags:
      - public
      summary: find land parcel by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ plot of land\n"
      operationId: parcela_wgs_api
      parameters:
      - name: lat
        in: query
        description: WGS-84 lat coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 49.0170822N
      - name: lon
        in: query
        description: WGS-84 lon coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 15.7719172E
      responses:
        "200":
          description: plot (of land) object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcela'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find plot (of land) by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ plot of land\n"
      operationId: parcela_wgs_post_api
      requestBody:
        description: WGS coordinates
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Wgs'
      responses:
        "200":
          description: plot (of land) object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcela'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/settlement/jtsk:
    get:
      tags:
      - public
      summary: find basic settlement unit by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: zsj_api
      parameters:
      - name: x
        in: query
        description: JTSK x coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -770559.237
      - name: "y"
        in: query
        description: JTSK y coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -1068317.864
      responses:
        "200":
          description: basic settlement unit object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zsj'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find basic settlement unit by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: zsj_post_api
      requestBody:
        $ref: '#/components/requestBodies/PointJtskReq'
      responses:
        "200":
          description: basic settlement unit object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zsj'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/settlement/wgs:
    get:
      tags:
      - public
      summary: find basic settlement unit by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: zsj_wgs_api
      parameters:
      - name: lat
        in: query
        description: WGS-84 lat coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 49.0170822N
      - name: lon
        in: query
        description: WGS-84 lon coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 15.7719172E
      responses:
        "200":
          description: basic settlement unit object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zsj'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find basic settlement unit by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: zsj_wgs_post_api
      requestBody:
        $ref: '#/components/requestBodies/PointWgsReq'
      responses:
        "200":
          description: basic settlement unit object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zsj'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/cadaster/jtsk:
    get:
      tags:
      - public
      summary: find cadastral territory by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ cadastral territory\n"
      operationId: ku_api
      parameters:
      - name: x
        in: query
        description: JTSK x coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -770559.237
      - name: "y"
        in: query
        description: JTSK y coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -1068317.864
      responses:
        "200":
          description: cadastral territory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KatastralniUzemi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find cadastral territory by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ cadastral territory\n"
      operationId: ku_post_api
      requestBody:
        $ref: '#/components/requestBodies/PointJtskReq'
      responses:
        "200":
          description: cadastral territory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KatastralniUzemi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/cadaster/wgs:
    get:
      tags:
      - public
      summary: find cadastral territory by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: ku_wgs_api
      parameters:
      - name: lat
        in: query
        description: WGS-84 lat coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 49.0170822N
      - name: lon
        in: query
        description: WGS-84 lon coordinate
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 15.7719172E
      responses:
        "200":
          description: cadastral territory object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KatastralniUzemi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find cadastral territory by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basic settlement unit\n"
      operationId: ku_wgs_post_api
      requestBody:
        $ref: '#/components/requestBodies/PointWgsReq'
      responses:
        "200":
          description: cadastral territory object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KatastralniUzemi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
  /reverse/basin/jtsk:
    get:
      tags:
      - public
      summary: find basin info by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basin info\n"
      operationId: povodi_api
      parameters:
      - name: x
        in: query
        description: JTSK x coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -770559.237
      - name: "y"
        in: query
        description: JTSK y coordinate
        required: true
        style: form
        explode: true
        schema:
          type: number
          format: double
          example: -1068317.864
      responses:
        "200":
          description: basin info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Povodi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
    post:
      tags:
      - public
      summary: find basin info by coordinates
      description: "By passing in the appropriate options, you can search for\nthe\
        \ basin info\n"
      operationId: povodi_post_api
      requestBody:
        $ref: '#/components/requestBodies/PointJtskReq'
      responses:
        "200":
          description: basin info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Povodi'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
...
  /convert/polygon/jtsk:
    post:
      tags:
      - public
      summary: converts polygon from WGS84 to JTSK
      description: "By passing in the appropriate options, you can obtain converted\
        \ value\n"
      operationId: convert_polygon_jtsk_api
      requestBody:
        $ref: '#/components/requestBodies/PolygonWgsReq'
      responses:
        "200":
          description: converted polygon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolygonJtsk'
        "204":
          description: no data returned
        "400":
          description: bad input parameter
        "401":
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
      security:
      - apiKey: []
      x-openapi-router-controller: swagger_server.controllers.public_controller
...
components:
  schemas:
...
    Jtsk:
      type: object
      properties:
        x:
          type: number
          format: double
          example: -773537.486
        "y":
          type: number
          format: double
          example: -1068166.65742
      description: JTSK coordinates
      example:
        x: -773537.486
        "y": -1068166.65742
...
    Parcela:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2554947605
        kmenovecislo:
          type: integer
          format: int64
          example: 92
        pododdelenicisla:
          type: integer
          format: int64
          example: 3
        vymeraparcely:
          type: number
          format: double
          example: 224
        administrative_division:
          $ref: '#/components/schemas/AdministrativeDivision'
      description: objekt parcely
      example:
        vymeraparcely: 224
        administrative_division:
          vusc_nazev: Královéhradecký kraj
          orp_nazev: Broumov
          regionsoudrznosti_kod: 51
          nuts_1: CZ
          nuts_lau1: CZ0523
          nuts_2: CZ05
          nuts_lau2: CZ0523547786
          nuts_3: CZ052
          obec_nazev: Adršpach
          pou_nazev: Teplice nad Metují
          ku_nazev: Dolní Adršpach
          obec_kod: 547786
          okres_nazev: Náchod
          okres_kod: 3605
          regionsoudrznosti_nazev: Severovýchod
          vusc_kod: 86
          spravni_obec_nazev: Broumov
          spravni_obec_kod: 573922
          obec_statuskod: 2
          pou_kod: 2283
          orp_kod: 1031
          ku_kod: 600059
        id: 2554947605
        kmenovecislo: 92
        pododdelenicisla: 3
...
  requestBodies:
...
    PointJtskReq:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PointJtsk'
...
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-KEY
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_apiKey